services/location-tracking/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-socketio==5.3.6
gunicorn==21.2.0
eventlet==0.36.1
psycopg2-binary==2.9.9
redis==5.0.7
PyJWT==2.8.0
python-dotenv==1.0.1
